+++ modules/task_1/pestreev_d_letter_count/letter_count.h
@@ -4,7 +4,7 @@
 #include <string>
 
 std::string getRandomString(int  size);
-int LetterInLineSequential(std::string str);
-int LetterInLineParallel(std::string global_str, int size_str);
+int64_t LetterInLineSequential(std::string str);
+int64_t LetterInLineParallel(std::string global_str, int size_str);
 
 #endif  // MODULES_TASK_1_PESTREEV_D_LETTER_COUNT_LETTER_COUNT_H_
